projects/tribes.ai/requirements.txt
contextlib2==21.6.0
neo4j-driver==4.3.1
neobolt==1.7.17
neomodel==4.0.5
python-dateutil==2.8.2
pytz==2021.1
schema==0.7.4
Shapely==1.7.1
six==1.16.0
